backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
httpx==0.27.2
gitpython==3.1.43
radon==6.0.1
coverage==7.6.4
pytest==8.3.3
flake8==7.1.1
requests==2.32.3